+++ blockservice/blockservice.go
@@ -13,6 +13,7 @@ import (
 
 	"github.com/ipfs/boxo/blockstore"
 	"github.com/ipfs/boxo/exchange"
+	"github.com/ipfs/boxo/provider"
 	"github.com/ipfs/boxo/verifcid"
 	blocks "github.com/ipfs/go-block-format"
 	"github.com/ipfs/go-cid"
@@ -73,10 +74,21 @@ type BoundedBlockService interface {
 
 var _ BoundedBlockService = (*blockService)(nil)
 
+// ProvidingBlockService is a Blockservice which provides new blocks to a provider.
+type ProvidingBlockService interface {
+	BlockService
+
+	// Provider can return nil, then no provider is used.
+	Provider() provider.Provider
+}
+
+var _ ProvidingBlockService = (*blockService)(nil)
+
 type blockService struct {
 	allowlist  verifcid.Allowlist
 	blockstore blockstore.Blockstore
 	exchange   exchange.Interface
+	provider   provider.Provider
 	// If checkFirst is true then first check that a block doesn't
 	// already exist to avoid republishing the block on the exchange.
 	checkFirst bool
@@ -96,6 +108,13 @@ func WriteThrough() Option {
 func WithAllowlist(allowlist verifcid.Allowlist) Option {
 	return func(bs *blockService) {
 		bs.allowlist = allowlist
+	}
+}
+
+// WithProvider allows to advertise anything that is added through the blockservice.
+func WithProvider(prov provider.Provider) Option {
+	return func(bs *blockService) {
+		bs.provider = prov
 	}
 }
 
@@ -133,6 +152,10 @@ func (s *blockService) Allowlist() verifcid.Allowlist {
 	return s.allowlist
 }
 
+func (s *blockService) Provider() provider.Provider {
+	return s.provider
+}
+
 // NewSession creates a new session that allows for
 // controlled exchange of wantlists to decrease the bandwidth overhead.
 // If the current exchange is a SessionExchange, a new exchange
@@ -168,6 +191,11 @@ func (s *blockService) AddBlock(ctx context.Context, o blocks.Block) error {
 	if s.exchange != nil {
 		if err := s.exchange.NotifyNewBlocks(ctx, o); err != nil {
 			logger.Errorf("NotifyNewBlocks: %s", err.Error())
+		}
+	}
+	if s.provider != nil {
+		if err := s.provider.Provide(o.Cid()); err != nil {
+			logger.Errorf("Provide: %s", err.Error())
 		}
 	}
 
@@ -216,6 +244,14 @@ func (s *blockService) AddBlocks(ctx context.Context, bs []blocks.Block) error {
 			logger.Errorf("NotifyNewBlocks: %s", err.Error())
 		}
 	}
+	if s.provider != nil {
+		for _, o := range toput {
+			if err := s.provider.Provide(o.Cid()); err != nil {
+				logger.Errorf("Provide: %s", err.Error())
+			}
+		}
+	}
+
 	return nil
 }
 
@@ -269,6 +305,12 @@ func getBlock(ctx context.Context, c cid.Cid, bs BlockService, fetchFactory func
 	}
 	if ex := bs.Exchange(); ex != nil {
 		err = ex.NotifyNewBlocks(ctx, blk)
+		if err != nil {
+			return nil, err
+		}
+	}
+	if provider := grabProviderFromBlockservice(bs); provider != nil {
+		err = provider.Provide(blk.Cid())
 		if err != nil {
 			return nil, err
 		}
@@ -346,6 +388,7 @@ func getBlocks(ctx context.Context, ks []cid.Cid, blockservice BlockService, fet
 		}
 
 		ex := blockservice.Exchange()
+		provider := grabProviderFromBlockservice(blockservice)
 		var cache [1]blocks.Block // preallocate once for all iterations
 		for {
 			var b blocks.Block
@@ -375,6 +418,14 @@ func getBlocks(ctx context.Context, ks []cid.Cid, blockservice BlockService, fet
 					return
 				}
 				cache[0] = nil // early gc
+			}
+
+			if provider != nil {
+				err = provider.Provide(b.Cid())
+				if err != nil {
+					logger.Errorf("could not tell the provider about new blocks: %s", err)
+					return
+				}
 			}
 
 			select {
@@ -462,4 +513,12 @@ func grabAllowlistFromBlockservice(bs BlockService) verifcid.Allowlist {
 		return bbs.Allowlist()
 	}
 	return verifcid.DefaultAllowlist
+}
+
+// grabProviderFromBlockservice can return nil if no provider is used.
+func grabProviderFromBlockservice(bs BlockService) provider.Provider {
+	if bbs, ok := bs.(ProvidingBlockService); ok {
+		return bbs.Provider()
+	}
+	return nil
 }